requirements.txt
discord.py>=2.3.2
asyncpg>=0.29.0
Pillow>=10.0.0
